Vaoious updates.

+++ FluxSolution/BaseLibrary/Source/Random/RandomUInt32/RandomUInt32.cs
@@ -40,7 +40,7 @@ namespace Flux.Random
 		/// <summary>Needs to return a value that is greater than or equal to 0.0, and less than 1.0</summary>
 		/// <returns>A double-precision floating point number that is greater than or equal to 0.0, and less than 1.0.</returns>
 		protected override double Sample()
-			=> (SampleUInt32() >> 8) / (1 << 24);
+			=> (double)(SampleUInt32() >> 8) / (double)(1 << 24);
 
 		internal abstract uint SampleUInt32();
 	}
